修改hotel.jsp,修改飯店model
修改hotel.jsp成正式的畫面
增加與hotel.jsp會用到的css和圖片
修改飯店套件中的3種DAO和VO

+++ iCastle/src/main/java/com/icastle/hotels/model/HotelDAOJDBC.java
@@ -133,7 +133,7 @@ public class HotelDAOJDBC implements HotelDAO_Interface {
 				hotel.setPw(rs.getString(4));
 				hotel.setAddr(rs.getString(5));
 				hotel.setZone(rs.getString(6));
-				hotel.setPoint(rs.getInt(7));
+				hotel.setPoint(rs.getDouble(7));
 				hotel.setHot(rs.getInt(8));
 				hotel.setStar(rs.getInt(9));
 				hotel.setHotelState(rs.getInt(10));
